requirements.txt
pick==2.4.0
PyAutoGUI==0.9.54
pynput==1.7.7
tweety_ns==2.1
